[#174902061] Refactoring Api call tasks
