+++ utils/api.ts
@@ -1,0 +1,42 @@
+import {
+  fromLeft,
+  TaskEither,
+  taskEither,
+  tryCatch
+} from "fp-ts/lib/TaskEither";
+import * as t from "io-ts";
+import { IResponseType } from "italia-ts-commons/lib/requests";
+import { ILogger } from "./logging";
+import {
+  ErrorResponses,
+  toDefaultResponseErrorInternal,
+  toErrorServerResponse
+} from "./responses";
+
+export const withApiRequestWrapper = <T>(
+  logger: ILogger,
+  apiCallWithParams: () => Promise<
+    t.Validation<IResponseType<number, T, never>>
+  >,
+  successStatusCode: number
+): TaskEither<ErrorResponses, T> =>
+  tryCatch(
+    () => apiCallWithParams(),
+    errs => {
+      logger.logUnknown(errs);
+      return toDefaultResponseErrorInternal(errs);
+    }
+  ).foldTaskEither(
+    err => fromLeft(err),
+    errorOrResponse =>
+      errorOrResponse.fold(
+        errs => {
+          logger.logErrors(errs);
+          return fromLeft(toDefaultResponseErrorInternal(errs));
+        },
+        responseType =>
+          responseType.status !== successStatusCode
+            ? fromLeft(toErrorServerResponse(responseType))
+            : taskEither.of(responseType.value)
+      )
+  );
